pyproject.toml
[project]
name = "pump_bot"
version = "2.0"
description = "Trade tokens on pump.fun"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "base58>=2.1.1",
    "borsh-construct>=0.1.0",
    "construct>=2.10.67",
    "construct-typing>=0.5.2",
    "solana==0.36.6",
    "solders>=0.26.0",
    "websockets>=15.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.11.13",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "protobuf>=5.29.4",
    "pyyaml>=6.0.2",
    "uvloop>=0.21.0; sys_platform != 'win32'",
    "prometheus_client>=0.20.0",
    "aiofiles>=23.2.1",
    "tenacity>=8.2.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
pump_bot = "bot_runner:main"
sell = "sell:main"
buy = "buy:main"

[tool.setuptools]
py-modules = ["buy", "sell", "bot_runner"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.ruff]
exclude = [
    ".git",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "__pycache__",
    "venv",
    "build",
    "dist",
    "learning-examples",
    "idl",
]
line-length = 100
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",
    "E402",
    "F401",
    "F841",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
